fix(auth, android): add browser dependency as crash workaround
The reCAPTCHA auth flow requires androidx.browser but it is not specified
as a transitive dependency upstream, so it crashes.

I attempted to use as wide a version range as I felt safe, and provided an
override if you specify "androidxBrowserVersion" in your android/build.gradle
ext config block

Related #4744
Related https://github.com/firebase/firebase-android-sdk/issues/2164

This may be removed when the upstream issue is closed.