feat(auth): Add multi-factor support for the sign-in flow (#6593)
* feat(auth): adds iOS/Web support for multi-factor sign-in flow

Adds code required to complete the sign-in flow for users that have enrolled
second factors. Due to a difference in the implementation of the
PhoneAuthProvider it is not possible to follow the implementation of the Web
API.

* docs(auth): update documentation for multi-factor authentication

Provide general setup and usage information for multi-factor authentication flows.

* fix(docs): Change prev link for Firestore docs to multi-factor auth

* feat(auth): Adds Android support for multi-factor sign-in flow

Implement the Android part required to support the multi-factor sign-in flow.
Makes the `multiFactor` property for the Firebase user object available as well.

Known issues:
- The enrollmentInfo for a MultiFactorInfo is currently reported with 0 on Android.

* feat(auth): multi-factor enroll feature for Android and e2e-tests

* feat(auth): Implement multi-factor enrollment for iOS

* fix(tests): More robust matchers for invalid-verification-code error

* Fix error message for invalid phone numbers

to match the error message produced by the Web.

* Add test and fix error message for unknown multi-factor hint

to match the error message produced by the Web.